refactor: Remove NotificationDispatcher from core
