+++ netwolf/lib/netwolf.dart
@@ -1,4 +1,4 @@
-export 'package:netwolf_core/netwolf_core.dart'
-    hide NotificationKey, NotificationName;
+export 'package:netwolf_core/netwolf_core.dart' hide BaseNetwolfController;
 
+export 'src/controller.dart';
 export 'src/ui/pages/netwolf_widget.dart' hide showCustomModalBottomSheet;
