+++ netwolf/lib/src/enums.dart
@@ -1,0 +1,9 @@
+enum NotificationName {
+  show,
+  search,
+  refetchRequests,
+  updateFilters,
+  clearFilters
+}
+
+enum NotificationKey { searchTerm, method, status }
