+++ netwolf/lib/src/ui/widgets/netwolf_search_bar.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:netwolf/src/enums.dart';
 import 'package:netwolf/src/ui/widgets/filter_dialog.dart';
 import 'package:netwolf_core/netwolf_core.dart';
 import 'package:notification_dispatcher/notification_dispatcher.dart';
