+++ netwolf_core/lib/netwolf_core.dart
@@ -1,5 +1,5 @@
 export 'src/core/enums.dart';
-export 'src/core/netwolf_controller.dart';
+export 'src/core/netwolf_controller.dart' hide MockNetwolfController;
 export 'src/core/typedefs.dart';
 export 'src/models/netwolf_request.dart';
 export 'src/models/result.dart';
