+++ netwolf_core/lib/src/core/enums.dart
@@ -47,13 +47,3 @@ enum HttpResponseStatus {
     return null;
   }
 }
-
-enum NotificationName {
-  show,
-  search,
-  refetchRequests,
-  updateFilters,
-  clearFilters
-}
-
-enum NotificationKey { searchTerm, method, status }
